Upgraded SignUp Page to be on a Whole New Level
1. Added validations to all input fields in SignUp
Page

---> I wrote all the validation logics from
scratch, because I wanted a customized error text
behaviour for my Chat App's SignUp form

---> I also added error texts to all input fields
in SignUp page and handled all the edge cases for
the inputs

2. Added a loading spinner to the SignUp Page

---> Whenever the user uploads a profile pic,
frontend takes some time to upload that image to
cloudinary, I already added a success or error
toast to this process when I was creating the
Sign Up page earlier. This toast pops-up to
tell the status of profile pic upload or deletion
whichever task the user is trying to do of, but
until that toast doesn't pops up, the user had
no idea earlier of what's going on behind the
scenes in the SignUp Page and the "Sign Up" button
was always active, so the user don't know that a
toast will appear to notify them about profile pic
successful upload and the name of the profile pic
is updated on the Profile Pic uploading button
instantaneously and the Sign Up button is active
all the time, and this leads to a very bad UI/UX
that's why to make the UI/UX better I added a
loading spinner to profile pics's uploading and
deletion process.

---> The loading spinner helps the user in knowing
that there is some process going-on behind the scenes,
and whenever the image upload or image deletion
compeletes as requested by user clicks, the loading
spinner disappears and a notifying toast appears for
telling the status of image upload or deletion

---> Overall I have also synced the "Sign Up"
button disabling and enabling with erros and the
laoding spinner. Whenever there is an error in any
of the input fields in SignUp page and whenever
the loading spinner is spinning, the "Sign Up"
button is disabled

3. I also updated MongoDB userModel and updated the
minLength of password to be 8 chars instead of being
6 chars

4. In SideBar component I also handled the width of
tooltip to be automatic for user's profile name, and
this helped in imporving the UI look better whenever
the user has too long name, this tooltip which I am
talking about right now is shown to user whenever
the user hovers on their profile pic

$ Yeahh, these are the only features that I added right
now, that's it for now.