Well, I am not able to resist thinking about new and new edge cases
--> One after another I am able to think of new
and new edge cases that how my frontend or backend
might fail

--> When I started to use my web app in a way that
I don't know anything about it, I ended up pressing
buttons in random way as a dumb user and I started
giving random inputs to input fields and like this
I am able to figure out new and new edge cases

--> Althought my web app is becoming robust as I
am follwing this practice of finding edge case and
then I am handling those edge cases, but like this
it is taking a lot of time even just to build a
sign up page, and I am not able to reach the zone
of building the core funcitonality of this web app
quickly (Obviously the core funcitonality of my
chat app is texting)

--> But I don't want to compromise on quality of
my projects, its okay even if it is taking time.
"Quality things take time in getting built."

--> And I am first time building a MERN stack
based web app with such high quality where I am
trying to handle all the edge cases and I am
trying to keep the code base readable, modular
and easy to manage and understand

@ Right now I handled few edge cases in Sign
Up page and let's see what were those edge cases:

1. The Sign-Up page's "Sign Up" button was
pressable even once it has been pressed, and
this was making the frontend to make redundant
calls to the backend if user presses the button
continously due to any random reason. So, I
handeld this case by making sure that once the
"Sign Up" button has been pressed it gets
disabled until the response for this first call
has been received from the backend

2. The profile pic upload button was also
pressable even after the pic was uploaded, or
even the profle pic was being uploaded or the
pic was being removed once user pressed the
cross button after uploading the pic. So, in
all these cases the profile pic upload button
should get disabled when either it is already
processing a request or it has already received
an input image. That's why to have a roubst web
application, I handled this edge case also by
disabling the input tag once it has the image
inputor when it is already processing a backend
call

3. Another edge case was that, there is a
<Link /> tag redirecting to Login page if user
clicks on it. So, if a user uploads a profile
pic and leaves the Sign Up process in between
and navigates to Login page then pic is still
uploaded to the cloudinary and it is making my
cloudinary account a digital graveyard, that's
why I handeled this case earlier by adding a
clean up funciton for this mid way navigation
of user to login page but I spotted the
navigation was also possible when the image
was being uploaded or deleted, and in this
situation the link tag should not allow the
user to redirect until that uploading/deletion
process of profile pic to/from cloudinary is
not completed. So, I handled this case also by
avoiding the navigation to happen when uploading
or deletion of profile pic is happening, I
prevented default behaviour of Link tag to occur
in this edge case for solving this issue

4. One more issue was with cross button which
appears for removing the pic after the user has
uploaded a profile pic, so even when the user
has pressed the sign Up button, that cross button
on image upload tag was active, and until backend
doesn't responds back, the UI of sign up page was
still accessible to user and if in this mean time
of 1ms if user presses that cross button then the
image deletion call will be made. So, I handled
this case by making sure that the crossbutton from
profile pic should get hidden when the sign up
button has been pressed and until we don't receive
a response from backend for this onClick action.